Switch map checking from random to PositionIteratorCircularSpiral.
This allows us to search an arc instead of a box or a full circle. Also now we iterate over positions instead of randomly searching for a set amount of time.

+++ screen/src/frame.rs
@@ -112,7 +112,7 @@ pub trait Frame {
         &self,
         fuzzy_pixel: FuzzyPixel,
         middle: Position,
-        radius: f32,
+        radius: i32,
     ) -> Option<Position> {
         let batch_size = 1000;
         let time = std::time::Instant::now();
